refactor(inbox): extract useInboxPlanApplyFlow from InboxPage
Move plan approval, apply (single/batch/all), cancel handlers, and
progress/busy state into a dedicated hook. InboxPage retains
composition; behavior unchanged.
